requirements.txt
discord.py
pytz
python-dotenv
openai
pymongo